Issue #36

+++ src/QuantitativeWorld.Conversion/LinearUnitConverter.cs
@@ -5,7 +5,7 @@ namespace QuantitativeWorld.Conversion
     public class LinearUnitConverter<TUnit> : IUnitConverter<TUnit>
         where TUnit : ILinearUnit
     {
-        public decimal ConvertValue(decimal value, TUnit sourceUnit, TUnit targetUnit) =>
+        public double ConvertValue(double value, TUnit sourceUnit, TUnit targetUnit) =>
             value * sourceUnit.ValueInBaseUnit / targetUnit.ValueInBaseUnit;
     }
 }
