+++ src/QuantitativeWorld/Angular/AngleFactory.cs
@@ -4,7 +4,7 @@ namespace QuantitativeWorld.Angular
 {
     public class AngleFactory : ILinearQuantityFactory<Angle, AngleUnit>
     {
-        public Angle Create(decimal value, AngleUnit unit) =>
+        public Angle Create(double value, AngleUnit unit) =>
             new Angle(value, unit);
     }
 }
