+++ src/QuantitativeWorld/Constants.cs
@@ -4,50 +4,50 @@ namespace QuantitativeWorld
 {
     static class Constants
     {
-        public const decimal KilogramsPerGram = 0.001m;
-        public const decimal KilogramsPerQuintal = 100m;
-        public const decimal KilogramsPerTon = 1000m;
-        public const decimal MilligramsPerGram = 1000m;
-        public const decimal GramsPerDecagram = 10m;
+        public const double KilogramsPerGram = 0.001d;
+        public const double KilogramsPerQuintal = 100d;
+        public const double KilogramsPerTon = 1000d;
+        public const double MilligramsPerGram = 1000d;
+        public const double GramsPerDecagram = 10d;
 
-        public const decimal KilogramsPerPound = 0.45359237m;
-        public const decimal OuncesPerPound = 16m;
-        public const decimal PoundsPerStone = 14m;
-        public const decimal PoundsPerCental = 100m;
-        public const decimal StonesPerHundredweight = 8m;
+        public const double KilogramsPerPound = 0.45359237d;
+        public const double OuncesPerPound = 16d;
+        public const double PoundsPerStone = 14d;
+        public const double PoundsPerCental = 100d;
+        public const double StonesPerHundredweight = 8d;
 
-        public const decimal MetresPerMillimetre = 0.001m;
-        public const decimal MetresPerCentimetre = 0.01m;
-        public const decimal MetresPerDecimetre = 0.1m;
-        public const decimal MetresPerKilometre = 1000m;
-        public const decimal MetresPerInch = 0.0254m;
-        public const decimal MetresPerFoot = 0.3048m;
-        public const decimal MetresPerYard = 0.9144m;
+        public const double MetresPerMillimetre = 0.001d;
+        public const double MetresPerCentimetre = 0.01d;
+        public const double MetresPerDecimetre = 0.1d;
+        public const double MetresPerKilometre = 1000d;
+        public const double MetresPerInch = 0.0254d;
+        public const double MetresPerFoot = 0.3048d;
+        public const double MetresPerYard = 0.9144d;
 
-        public const decimal MetresPerNauticalMile = 1852m;
+        public const double MetresPerNauticalMile = 1852d;
 
-        public const decimal InchesPerFoot = 12m;
-        public const decimal FeetPerYard = 3m;
-        public const decimal YardsPerRod = 5.5m;
-        public const decimal RodsPerChain = 4m;
-        public const decimal ChainsPerFurlong = 10m;
-        public const decimal FurlongsPerMile = 8m;
+        public const double InchesPerFoot = 12d;
+        public const double FeetPerYard = 3d;
+        public const double YardsPerRod = 5.5d;
+        public const double RodsPerChain = 4d;
+        public const double ChainsPerFurlong = 10d;
+        public const double FurlongsPerMile = 8d;
 
-        public const decimal CablesPerNauticalMile = 10m;
+        public const double CablesPerNauticalMile = 10d;
 
-        public const decimal WattsPerMilliwatt = 0.001m;
-        public const decimal WattsPerKilowatt = 1000m;
-        public const decimal WattsPerMegawatt = 1000000m;
-        public const decimal WattsPerMechanicalHorsepower = 76.0402249m * 9.80665m;
-        public const decimal WattsPerMetricHorsepower = 75m * 9.80665m;
-        public const decimal WattsPerErgPerSecond = 0.0000001m;
-        public const decimal WattsPerStheneMetrePerSecond = 1000m;
+        public const double WattsPerMilliwatt = 0.001d;
+        public const double WattsPerKilowatt = 1000d;
+        public const double WattsPerMegawatt = 1000000d;
+        public const double WattsPerMechanicalHorsepower = 76.0402249d * 9.80665d;
+        public const double WattsPerMetricHorsepower = 75d * 9.80665d;
+        public const double WattsPerErgPerSecond = 0.0000001d;
+        public const double WattsPerStheneMetrePerSecond = 1000d;
 
-        public const decimal RadiansPerTurn = (decimal)Math.PI;
-        public const decimal DegreesPerTurn = 360m;
-        public const decimal ArcminutesPerTurn = DegreesPerTurn * 60m;
-        public const decimal ArcsecondsPerTurn = ArcminutesPerTurn * 60m;
-        public const decimal GradiansPerTurn = 400m;
-        public const decimal NATOMilsPerTurn = 6400m;
+        public const double RadiansPerTurn = Math.PI;
+        public const double DegreesPerTurn = 360d;
+        public const double ArcminutesPerTurn = DegreesPerTurn * 60d;
+        public const double ArcsecondsPerTurn = ArcminutesPerTurn * 60d;
+        public const double GradiansPerTurn = 400d;
+        public const double NATOMilsPerTurn = 6400d;
     }
 }
