+++ src/QuantitativeWorld/LengthUnit.cs
@@ -7,13 +7,13 @@ namespace QuantitativeWorld
     {
         private readonly string _name;
         private readonly string _abbreviation;
-        private readonly decimal? _valueInMetres;
+        private readonly double? _valueInMetres;
 
-        public LengthUnit(string name, string abbreviation, decimal valueInMetres)
+        public LengthUnit(string name, string abbreviation, double valueInMetres)
         {
             Assert.IsNotNullOrWhiteSpace(name, nameof(name));
             Assert.IsNotNullOrWhiteSpace(abbreviation, nameof(abbreviation));
-            Assert.IsGreaterThan(valueInMetres, 0m, nameof(valueInMetres));
+            Assert.IsGreaterThan(valueInMetres, 0d, nameof(valueInMetres));
 
             _name = name;
             _abbreviation = abbreviation;
@@ -22,10 +22,10 @@ namespace QuantitativeWorld
 
         public string Name => _name ?? Metre._name;
         public string Abbreviation => _abbreviation ?? Metre._abbreviation;
-        public decimal ValueInMetres => _valueInMetres ?? Metre._valueInMetres.Value;
+        public double ValueInMetres => _valueInMetres ?? Metre._valueInMetres.Value;
 
         public override string ToString() => Abbreviation;
 
-        decimal ILinearUnit.ValueInBaseUnit => ValueInMetres;
+        double ILinearUnit.ValueInBaseUnit => ValueInMetres;
     }
 }
